// packages/entities/entities-plugins/src/components/free-form/plugins/datakit/types.ts
export type NodePhase = 'request' | 'response'

/************************************************
 *             Editor global store              *
 ************************************************/

export type NodeId = `node:${number}`
export type EdgeId = `edge:${number}`
export type FieldId = `field:${number}`
export type GroupId = `${NodeId}:${BranchName}`
// ...
export interface EdgeData {
  source: NodeId
  sourceField?: FieldId
  target: NodeId
  targetField?: FieldId
}

export interface EdgeInstance extends EdgeData {
  id: EdgeId
}

export interface GroupInstance {
  id: GroupId
  ownerId: NodeId
  branch: BranchName
  phase: NodePhase
  memberIds: NodeId[]
  position?: XYPosition
  dimensions?: Dimensions
}